# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NCDMB Corporative API",
        "version": "1.0.0"
    },
    "paths": {
        "/agents": {
            "get": {
                "tags": [
                    "Agents"
                ],
                "summary": "Returns all agents on the system",
                "description": "Returns all agents on the system",
                "operationId": "findRoles",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Agent"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Page Not Found. If error persists, contact info@ncdmb.gov.ng"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Error, please fix the following error(s)!;",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Agents"
                ],
                "summary": "Save Agent",
                "operationId": "agents",
                "parameters": [
                    {
                        "name": "code",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "label",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "path",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "short_name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "inReview",
                                "verified"
                            ]
                        }
                    },
                    {
                        "name": "isActive",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Agent   has been created successfully!",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Page Not Found. If error persists, contact info@ncdmb.gov.ng"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Error, please fix the following error(s)!;",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/agents/{id}": {
            "get": {
                "tags": [
                    "Agents"
                ],
                "summary": "Get specification by id",
                "description": "Returns based on id ",
                "operationId": "showRole",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "specification id to get",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Agent for  details!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Agent"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Page Not Found. If error persists, contact info@ncdmb.gov.ng"
                    },
                    "500": {
                        "description": "Error, please fix the following error(s)!;",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "This ID is invalid"
                    }
                }
            },
            "put": {
                "tags": [
                    "Agents"
                ],
                "summary": "update specification by database",
                "description": "Updates specification in database",
                "operationId": "updateRole",
                "parameters": [
                    {
                        "name": "code",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "label",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "short_name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "isActive",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Error, please fix the following error(s)!;",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Page Not Found. If error persists, contact info@ncdmb.gov.ng"
                    },
                    "422": {
                        "description": "Invalid specification id"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Agents"
                ],
                "summary": "remove specification from database",
                "description": "Deletes specification in database",
                "operationId": "updateRole",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "specification id to delete",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Agent deleted successfully!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Agent"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error, please fix the following error(s)!;",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Page Not Found. If error persists, contact info@ncdmb.gov.ng"
                    },
                    "422": {
                        "description": "Invalid specification id"
                    }
                }
            }
        },
        "/agents/{id}/edit": {
            "get": {
                "tags": [
                    "Agents"
                ],
                "summary": "Open form to edit specification",
                "description": "Returns based on id ",
                "operationId": "editRole",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "specification id to edit",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Agent"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error, please fix the following error(s)!;",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Page Not Found. If error persists, contact info@ncdmb.gov.ng"
                    },
                    "422": {
                        "description": "Invalid specification id"
                    }
                }
            }
        },
        "/contributions": {
            "get": {
                "tags": [
                    "Contributions"
                ],
                "summary": "Returns all contributions on the system",
                "description": "Returns all contributions on the system",
                "operationId": "findRoles",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Contribution"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Page Not Found. If error persists, contact info@ncdmb.gov.ng"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Error, please fix the following error(s)!;",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Contributions"
                ],
                "summary": "Save  contribution",
                "operationId": "contributions",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "fee",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    },
                    {
                        "name": "month",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "current",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "previous",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "json"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Contribution   has been created successfully!",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Page Not Found. If error persists, contact info@ncdmb.gov.ng"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Error, please fix the following error(s)!;",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/contributions/{id}/edit": {
            "get": {
                "tags": [
                    "Contributions"
                ],
                "summary": "Open form to edit contribution",
                "description": "Returns based on id ",
                "operationId": "editRole",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "contribution id to edit",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Contribution"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error, please fix the following error(s)!;",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Page Not Found. If error persists, contact info@ncdmb.gov.ng"
                    },
                    "422": {
                        "description": "Invalid contribution id"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Agent": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "code": {
                        "type": "string",
                        "example": "Agent Code"
                    },
                    "name": {
                        "type": "string",
                        "example": "Agent Name"
                    },
                    "label": {
                        "type": "string",
                        "example": "Agent Label"
                    },
                    "path": {
                        "type": "string",
                        "example": "Storage Path"
                    },
                    "short_name": {
                        "type": "string",
                        "example": "Agent Short Name"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "inReview",
                            "verified"
                        ],
                        "example": "verified"
                    },
                    "isActive": {
                        "type": "boolean",
                        "example": "True"
                    },
                    "created_at": {
                        "type": "date",
                        "example": "2020-10-20"
                    },
                    "updated_at": {
                        "type": "date",
                        "example": "2020-12-22"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Agent"
                }
            },
...
        }
    },
    "security": [
        []
    ]
}